% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cdr3_anchors.R
\name{add_cdr3_anchors}
\alias{add_cdr3_anchors}
\title{Add CDR3 anchor residues}
\usage{
add_cdr3_anchors(sequence_df, chains, species = "human")
}
\arguments{
\item{sequence_df}{A data.frame containing CDR3 sequences and J genes. Columns
must be named cdr3_beta and j_beta and/or cdr3_alpha and j_alpha.}

\item{chains}{Which chains to process. "B" for beta chain only, "A" for alpha,
"AB" for both.}

\item{species}{From which species are the sequences. Currently only human and
mouse are supported.}
}
\value{
Same data.frame as input with additional columns junction_beta and/or
jucntion_alpha
}
\description{
Construct full TCR junction sequence by adding conservative
anchor residues (C and F/W) to CDR3 sequence.
}
\examples{
# make example table without anchor residues
df <- example_TCR_df
df$cdr3_beta <- substr(df$junction_beta, 2, nchar(df$junction_beta) - 2)

# generate column with beta chain junction sequence
df <- add_cdr3_anchors(df, "B", species="human")
}
