% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterize_TCR.R
\name{clusterize_TCR}
\alias{clusterize_TCR}
\title{Find TCR clusters}
\usage{
clusterize_TCR(
  sequence_df,
  chains,
  tmp_folder,
  id_col,
  scores_filename = NA,
  threshold = NA,
  ncores = 1
)
}
\arguments{
\item{sequence_df}{A data.frame containing TCR sequence data. Each row must describe a unique TCR sequence. Required columns: cdr3_beta, v_beta, j_beta (for beta chain or paired chains clustering) and cdr3_alpha, v_alpha, j_alpha (paired chains clustering only).}

\item{chains}{"B" for beta chain only, "AB" for paired chains.}

\item{tmp_folder}{Path to a directory for storing temporary files which are deleted after clustering is finished.}

\item{id_col}{Name of a column with unique ids for each TCR (optional).}

\item{scores_filename}{If character string for naming a file is provided, BL-scores of each TCR pair will be exported to this file. Supported formats: .Rds, .csv.}

\item{threshold}{Clustering threshold (optional).}

\item{ncores}{The number of cores to use for parallel computation (default = 1).}
}
\value{
A data.frame containing same information as sequence_df plus the cluster ids. If scores_filename is provided a file with pairwise BL-scores is created.
}
\description{
Find TCR clusters
}
\examples{
'clusterize_TCR(sequence_df, chains="AB, tmp_folder=".", ncores=4)'
}
